requirements.txt
requests==2.32.3
duckdb>=0.10.4
dbt-duckdb==1.8.1
pytest==8.2.2
responses==0.25.3
python-dotenv==1.0.1
